// src/utils/api/api.ts
export interface Contract {
    /**
     * ID of the contract.
     * @type {string}
     * @memberof Contract
     */
    'id': string;
    /**
     * The symbol of the faction that this contract is for.
     * @type {string}
     * @memberof Contract
     */
    'factionSymbol': string;
    /**
     * Type of contract.
     * @type {string}
     * @memberof Contract
     */
    'type': ContractTypeEnum;
    /**
     * 
     * @type {ContractTerms}
     * @memberof Contract
     */
    'terms': ContractTerms;
    /**
     * Whether the contract has been accepted by the agent
     * @type {boolean}
     * @memberof Contract
     */
    'accepted': boolean;
    /**
     * Whether the contract has been fulfilled
     * @type {boolean}
     * @memberof Contract
     */
    'fulfilled': boolean;
    /**
     * Deprecated in favor of deadlineToAccept
     * @type {string}
     * @memberof Contract
     * @deprecated
     */
    'expiration': string;
    /**
     * The time at which the contract is no longer available to be accepted
     * @type {string}
     * @memberof Contract
     */
    'deadlineToAccept'?: string;
}

export const ContractTypeEnum = {
    Procurement: 'PROCUREMENT',
    Transport: 'TRANSPORT',
    Shuttle: 'SHUTTLE'
} as const;

export type ContractTypeEnum = typeof ContractTypeEnum[keyof typeof ContractTypeEnum];

/**
 * The details of a delivery contract. Includes the type of good, units needed, and the destination.
 * @export
 * @interface ContractDeliverGood
 */
export interface ContractDeliverGood {
    /**
     * The symbol of the trade good to deliver.
     * @type {string}
     * @memberof ContractDeliverGood
     */
    'tradeSymbol': string;
    /**
     * The destination where goods need to be delivered.
     * @type {string}
     * @memberof ContractDeliverGood
     */
    'destinationSymbol': string;
    /**
     * The number of units that need to be delivered on this contract.
     * @type {number}
     * @memberof ContractDeliverGood
     */
    'unitsRequired': number;
    /**
     * The number of units fulfilled on this contract.
     * @type {number}
     * @memberof ContractDeliverGood
     */
    'unitsFulfilled': number;
}
/**
 * Payments for the contract.
 * @export
 * @interface ContractPayment
 */
export interface ContractPayment {
    /**
     * The amount of credits received up front for accepting the contract.
     * @type {number}
     * @memberof ContractPayment
     */
    'onAccepted': number;
    /**
     * The amount of credits received when the contract is fulfilled.
     * @type {number}
     * @memberof ContractPayment
     */
    'onFulfilled': number;
}
/**
 * The terms to fulfill the contract.
 * @export
 * @interface ContractTerms
 */
export interface ContractTerms {
    /**
     * The deadline for the contract.
     * @type {string}
     * @memberof ContractTerms
     */
    'deadline': string;
    /**
     * 
     * @type {ContractPayment}
     * @memberof ContractTerms
     */
    'payment': ContractPayment;
    /**
     * The cargo that needs to be delivered to fulfill the contract.
     * @type {Array<ContractDeliverGood>}
     * @memberof ContractTerms
     */
    'deliver'?: Array<ContractDeliverGood>;
}